feat: ФЗ Спринт 14

+++ src/main/java/ru/practicum/shareit/item/ItemStorage.java
@@ -8,7 +8,7 @@ public interface ItemStorage {
 
     Item create(Item item);
 
-    Item findItemById(int id);
+    Item findItemById(Long id);
 
     Item update(Item item);
